Some (non-NCEP) grib messages have an illegal local-table flag, i.e. they report localtab=0 even if some parameter category, discipline or number value have a setting > 192 which indicates "Reserved for local use". This results in error messages like "**** ERROR: local table = 0 is not allowed, set to 1 ***".
Up to now, wgrib2 forcibly set localtab=1 in such cases and looks for
gribtable entries with localtab=1.

However, when using the wgrib2 option "-gribtable_used" on such a grib
message, an illegal gribtable is produced which can not be used as a
custom, user-defined gribtable (by setting "export GRIB2TABLE=...").

As "-gribtable_used" is both used to dump/document parameter values as
found in the grib message and as a template for a custom user gribtable,
the illegal state of the table is retained.
Instead, the table lookup is relaxed by adding a fallback pass with
corrected localtab setting. In the first pass, an entry is searched having
the original, maybe illegal, localtab setting. If no entry is found, then a
second search pass is made looking for localtab=1 entries.

Tone down the error message to a warning message.